Enable strictNullChecks for tests
By convention established in the DefinitelyTyped repo, types are written for `strictNullChecks: true`.
Disabling this option may cause breakage where types are accidentally resolved to optional or `undefined` instead of required due to `null` and `undefined` being ignored by the type-checker.

We already have this option enabled in other projects which leads to inconsistent types in the actual packages and how we consume them in tests.

It does make sense to have looser type-checking in tests but TypeScript has to apply a config to all files in the compilation not just a subset.

There's probably a larger discussion to be had what we really want out of having tests written in TypeScript.
Right now we can just follow the "write tests like our users would" mentality.

Opening this early to highlight existing breakages that need fixing.