Escape url-unsafe characters in names of app router scripts and styles
Building on #58293, this expands escaping of url-unsafe characters in paths to “required” scripts and styles in the app renderer.

This also refactors the test introduced in #58293 and expands it to include stylesheet references as well as checking resources in the head, which include special characters like turbopack references like `[turbopack]`.

Test Plan: `TURBOPACK=1 pnpm test-dev test/e2e/app-dir/resource-url-encoding`